feat: takes Switch component into it's own seperate file
