+++ index.js
@@ -2,32 +2,3 @@ import React, {
 	Component,
 	Children,
 } from 'react';
-
-export class Switch extends Component {	
-	render() {
-		const props  = this.props;
-		const { $, target, children = [] } = props;
-		const Cases = React.Children.map(
-			children,
-			child => React.cloneElement(child, { target: target })
-		);
-		return (
-			<$ {...props}>
-				{Cases}
-			</$>
-		);
-	}
-
-	static Case(props) {
-		const { $, value, target, children } = props;
-		if(target === value) {
-			return (
-				<$ {...props}>
-					{children}
-				</$>
-			);
-		} else {
-			return null;
-		}
-	}
-}
